fix: create a tools module for independent functions
