+++ packages/hashes.csv
@@ -8,4 +8,4 @@ valory/skills/abstract_abci,QmQeyVZThnkfxmXkUjBa1T55Kixssz3Lv82upCH5wHtayb
 valory/skills/abstract_round_abci,QmVSzSFKHQUQTkKjubyN4MicW5BK2MjpR2D9cNr5dNUAmm
 valory/skills/counter,QmSmt61iahp11PaYT6oNEzijMU8n15b1A4fty6HXmxi2gW
 valory/skills/counter_client,QmYcfMC1nzgeMMys5NHUXJTB3KKxwJFBM4tryBMpNggEHt
-valory/skills/price_estimation_abci,QmSXpDsrqpymqs3meZ8PpUaL15FuFGaVbe91oHNMPHqJB9
+valory/skills/price_estimation_abci,QmexxkTzbG1ADjZdT6yQ5YM5fv9Mj2cmSi5cAfePV2DmGt
