+++ packages/valory/skills/price_estimation_abci/skill.yaml
@@ -9,15 +9,15 @@ aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   README.md: QmRoHvdbCkafxVGBcZVM7oPQALSpfURbszxMKgwJuTbwjF
   __init__.py: QmaSFhnM257jBvNorhGMKEwtda6Jf1UzWyFAAQ47dsz9yJ
-  behaviours.py: QmQYsr3MMReLNnvgBGSe67MKqahhALEvjDAbMKvEBkhtcq
+  behaviours.py: QmeyfznYJ2CWCDaMfHLaxVfd7h18p4Qw4Rfa2AYKVb938N
   dialogues.py: QmdBmUMWoPbS2faqn8LicQVJfN6a2WjtdZDWFbiYoy7hfY
-  estimator.py: QmdxC8cozUky7T3nNa2rAKFB5dK4GH5xCfFRyGjypz9ZuC
   handlers.py: QmSLoyFCRkCW5iXpHzwfoUduANi9JU9BoRUghwBhbUnRcP
   models.py: QmUGP65hhDp8w1BwxfWa4NGUzE78rC8bsXnJqncDaAxqjf
   params.py: QmYiRjBiVXb5VLVmsR4RDWxcfmVGRW4b3Q8iyQXV29iNe5
   payloads.py: QmY9rJt3VaQ294yJzaBbA7BMtdophW8wdghwNP44PfELyi
   price_api.py: QmTvGaH16idZdFybvNw8G536S1Kvz4xdDFY8zvTdkBy8v6
-  rounds.py: QmbXf8F3HN3LKXthPAuWCCmVDvqkRbBg7MoHiLM7ci7JtB
+  rounds.py: Qmd32mFewjUdGga969Yer6HK9sBM2pVqoNfamRXqLDdB1k
+  tools.py: QmccW1Gtt9iZvE4tWF5EQveH8ZsLrUvZaajr7XEUBoSu5F
 fingerprint_ignore_patterns: []
 connections:
 - fetchai/http_client:0.22.0
